[rigging] Cleanup pass: collapse subpackages, trim docstrings, harden URL parser
Folds rigging.auth/ and rigging.resolver/ into single-file modules, collapses
the four resolver test files into one, drops over-protective lazy imports for
httpx/google-auth (now hard runtime deps), removes verbose docstrings and
__all__ blocks, deletes test-only `is_revoked`, rejects userinfo/port in
ServiceURL.parse, and replaces the iris plugin's manual close() with a context
manager. Net -715 lines of source and test code.
