+++ lib/rigging/pyproject.toml
@@ -9,13 +9,14 @@ requires-python = ">=3.11,<3.13"
 dependencies = [
     "fsspec>=2024.0.0",
     "gcsfs>=2024.0.0",
+    "google-auth>=2.0",
+    "httpx>=0.28.1",
     "PyJWT>=2.12.0",
     "s3fs>=2024.0.0",
 ]
 
 [dependency-groups]
 dev = [
-    "httpx>=0.28.1",
     "pytest>=8.3.2",
     "pytest-timeout",
 ]
