--- lib/rigging/src/rigging/auth/__init__.py
@@ -1,8 +1,0 @@
-# Copyright The Marin Authors
-# SPDX-License-Identifier: Apache-2.0
-
-"""Stateless authentication primitives shared across Marin services."""
-
-from rigging.auth.verifier import JWT_ALGORITHM, JwtVerifier, VerifiedIdentity
-
-__all__ = ["JWT_ALGORITHM", "JwtVerifier", "VerifiedIdentity"]
