--- lib/rigging/tests/resolver/__init__.py
@@ -1,2 +1,0 @@
-# Copyright The Marin Authors
-# SPDX-License-Identifier: Apache-2.0
